+++ Client-Server-API/go.mod
@@ -2,4 +2,4 @@ module github.com/bianavic/fullcycle_desafios.git
 
 go 1.23.0
 
-require github.com/mattn/go-sqlite3 v1.14.24 // indirect
+require github.com/mattn/go-sqlite3 v1.14.24
